requirements.txt
python-openstacksdk